requirements.txt
astroid==2.2.5
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Django==2.2
django-environ==0.4.5
djangorestframework==3.9.2
ffmpy==0.2.2
google-api-core==1.9.0
google-auth==1.6.3
google-cloud-speech==1.0.0
googleapis-common-protos==1.5.9
grpcio==1.19.0
idna==2.8
isort==4.3.17
lazy-object-proxy==1.3.1
mccabe==0.6.1
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pylint==2.3.1
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.1
urllib3==1.24.1
wrapt==1.11.1
